# Add configurable throughput parameters for clients to API server